Rename block-label to multiple-choice
`block-label` made sense when the labels were presented as buttons in an effort to get people to click on them. If we’re updating the markup in order to not require JS then we can take the opportunity to rename them to better match the functionality they expose.

govuk_frontend_toolkit is (annoyingly) coupled to `.block-label` (https://github.com/alphagov/govuk_frontend_toolkit/blob/e03e7f56e1b70194310ee970a3ccf00bd9376b36/javascripts/govuk/show-hide-content.js#L13) and would need to be updated as well if this lands.