Fix click selection in IE9 and 10
IE9/10 will happily change state when a click event is triggered against the label text, the space for the bullet, or outside the radio border but within the overall area. They will not change state if a user clicks in the gap inside the border and outside the bullet area (although focus state changes).

Giving the transparent input an explicit z-index causes IE9/10 to recognise it as a target and assign the selection correctly.